[python] Move python_lambda into dedicated lambda/ folder
Follows the existing single-concern-module convention already used
by set/, tuple/, math/, and module/ under src/python-frontend/.
Pure file relocation plus include-path/CMakeLists updates, no logic
changes.
