+++ src/python-frontend/converter/converter_expr.cpp
@@ -9,7 +9,7 @@
 #include <python-frontend/exception/python_exception_handler.h>
 #include <python-frontend/python_expr_builder.h>
 #include <python-frontend/math/python_int_overflow.h>
-#include <python-frontend/python_lambda.h>
+#include <python-frontend/lambda/python_lambda.h>
 #include <python-frontend/python-list/python_list.h>
 #include <python-frontend/math/python_math.h>
 #include <python-frontend/string/string_builder.h>
