+++ src/python-frontend/lambda/python_lambda.cpp
@@ -1,4 +1,4 @@
-#include <python-frontend/python_lambda.h>
+#include <python-frontend/lambda/python_lambda.h>
 #include <python-frontend/python_converter.h>
 #include <python-frontend/python_expr_builder.h>
 #include <python-frontend/type/type_handler.h>
