+++ src/python-frontend/python_converter.cpp
@@ -13,7 +13,7 @@
 #include <python-frontend/python_converter.h>
 #include <python-frontend/python-dict/python_dict_handler.h>
 #include <python-frontend/exception/python_exception_handler.h>
-#include <python-frontend/python_lambda.h>
+#include <python-frontend/lambda/python_lambda.h>
 #include <python-frontend/python-list/python_list.h>
 #include <python-frontend/type/python_typechecking.h>
 #include <python-frontend/string/string_builder.h>
